pyproject.toml
[project]
name = "agentes-busca-dinamica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai-tools>=0.36.0",
    "crewai>=0.100.1",
    "python-dotenv>=1.0.1",
    "elevenlabs>=1.51.0",
    "whisper-openai>=1.0.0",
    "evolutionapi>=0.0.9",
    "flask>=3.1.0",
]
